#487 Move out card swipe functionality into a single re-usable component
This is for a re-use in Landing page for new swipe pattern
